# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RealEstate AI API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /zones:
    get:
      operationId: listZones
      tags: [zones]
      summary: List all zones with market indicators
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Search filter
      responses:
        "200":
          description: Zones list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Zone"
  /zones/{id}:
    get:
      operationId: getZone
      tags: [zones]
      summary: Get a zone by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Zone
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Zone"

  # ── Zone Notes ─────────────────────────────────────────────────────────────
...
  /zone-notes/{zoneId}:
    get:
      operationId: getZoneNote
      tags: [zoneNotes]
      summary: Get note for a specific zone
      parameters:
        - name: zoneId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Zone note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneNote"
        "404":
          description: No note for this zone
    put:
      operationId: upsertZoneNote
      tags: [zoneNotes]
      summary: Create or update note for a zone
      parameters:
        - name: zoneId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZoneNoteInput"
      responses:
        "200":
          description: Saved zone note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoneNote"

  # ── AI Chat ────────────────────────────────────────────────────────────────
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadUrlRequest"
      responses:
        "200":
          description: Presigned upload URL generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadUrlResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /storage/objects/{filePath}:
    get:
      tags: [Storage]
      operationId: getObject
      summary: Serve a private object
      parameters:
        - name: filePath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object content
          content:
            "*/*":
              schema:
                type: string
                format: binary
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Users ────────────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        email:
          type: string
        fullName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [user, admin]
        investorType:
          type: ["string", "null"]
          enum: [fix_flip, rental, development, mixed, null]
        bio:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, clerkId, email, role, createdAt]
    UserUpdate:
      type: object
      properties:
        fullName:
          type: string
        avatarUrl:
          type: string
        investorType:
          type: string
          enum: [fix_flip, rental, development, mixed]
        bio:
          type: string
...
    Zone:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
        pricePerM2:
          type: number
        annualGrowthPct:
          type: number
        rentalYieldPct:
          type: number
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, name, city, pricePerM2, annualGrowthPct, rentalYieldPct, createdAt]
    ZoneNote:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        zoneId:
          type: integer
        content:
          type: string
        updatedAt:
          type: string
          format: date-time
        zone:
          $ref: "#/components/schemas/Zone"
      required: [id, userId, zoneId, content, updatedAt, zone]
    ZoneNoteInput:
      type: object
      properties:
        content:
          type: string
      required: [content]

    # ── AI Chat ───────────────────────────────────────────────────────────────
...
    UploadUrlRequest:
      type: object
      required: [fileName, contentType]
      properties:
        fileName:
          type: string
        contentType:
          type: string
        size:
          type: integer
    UploadUrlResponse:
      type: object
      required: [uploadUrl, objectPath]
      properties:
        uploadUrl:
          type: string
        objectPath:
          type: string